+++ src/main/java/mil/nga/geopackage/factory/GeoPackageCoreImpl.java
@@ -21,10 +21,13 @@ import mil.nga.geopackage.db.GeoPackageCoreConnection;
 import mil.nga.geopackage.db.GeoPackageTableCreator;
 import mil.nga.geopackage.extension.Extensions;
 import mil.nga.geopackage.extension.ExtensionsDao;
+import mil.nga.geopackage.extension.NGAExtensions;
 import mil.nga.geopackage.extension.index.GeometryIndex;
 import mil.nga.geopackage.extension.index.GeometryIndexDao;
 import mil.nga.geopackage.extension.index.TableIndex;
 import mil.nga.geopackage.extension.index.TableIndexDao;
+import mil.nga.geopackage.extension.link.FeatureTileLink;
+import mil.nga.geopackage.extension.link.FeatureTileLinkDao;
 import mil.nga.geopackage.features.columns.GeometryColumns;
 import mil.nga.geopackage.features.columns.GeometryColumnsDao;
 import mil.nga.geopackage.features.columns.GeometryColumnsSfSql;
@@ -702,6 +705,8 @@ public abstract class GeoPackageCoreImpl implements GeoPackageCore {
 	public void deleteTable(String table) {
 		verifyWritable();
 
+		NGAExtensions.deleteTableExtensions(this, table);
+		
 		ContentsDao contentsDao = getContentsDao();
 		contentsDao.deleteTable(table);
 	}
@@ -805,6 +810,14 @@ public abstract class GeoPackageCoreImpl implements GeoPackageCore {
 	 * {@inheritDoc}
 	 */
 	@Override
+	public void dropTable(String table) {
+		tableCreator.dropTable(table);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
 	public TableIndexDao getTableIndexDao() {
 		return createDao(TableIndex.class);
 	}
@@ -854,6 +867,35 @@ public abstract class GeoPackageCoreImpl implements GeoPackageCore {
 		} catch (SQLException e) {
 			throw new GeoPackageException("Failed to check if "
 					+ GeometryIndex.class.getSimpleName()
+					+ " table exists and create it", e);
+		}
+		return created;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public FeatureTileLinkDao getFeatureTileLinkDao() {
+		return createDao(FeatureTileLink.class);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public boolean createFeatureTileLinkTable() {
+		verifyWritable();
+
+		boolean created = false;
+		FeatureTileLinkDao dao = getFeatureTileLinkDao();
+		try {
+			if (!dao.isTableExists()) {
+				created = tableCreator.createFeatureTileLink() > 0;
+			}
+		} catch (SQLException e) {
+			throw new GeoPackageException("Failed to check if "
+					+ FeatureTileLink.class.getSimpleName()
 					+ " table exists and create it", e);
 		}
 		return created;